Refactor Supabase utilities into modular structure with services and types
