+++ utils/supabase/client.ts
@@ -1,0 +1,67 @@
+import { createClient } from "@supabase/supabase-js";
+import { configureSynced } from "@legendapp/state/sync";
+import { syncedSupabase } from "@legendapp/state/sync-plugins/supabase";
+import { observablePersistAsyncStorage } from "@legendapp/state/persist-plugins/async-storage";
+import { AppState, Platform } from "react-native";
+import "react-native-get-random-values";
+import "react-native-url-polyfill/auto";
+import { v4 as uuidv4 } from "uuid";
+import "whatwg-fetch";
+
+import { Database } from "@/utils/database.types";
+import { Storage } from "@/utils/storage";
+import { supabaseStorage } from "./storage";
+
+// =====================================================
+// CONFIGURATION SUPABASE CLIENT
+// =====================================================
+
+export const supabase = createClient<Database>(
+  process.env.EXPO_PUBLIC_SUPABASE_URL || "http://localhost:54321",
+  process.env.EXPO_PUBLIC_SUPABASE_ANON_KEY || "your-anon-key",
+  {
+    auth: {
+      // Utiliser notre storage universel
+      ...(supabaseStorage ? { storage: supabaseStorage } : {}),
+      autoRefreshToken: true,
+      persistSession: true,
+      detectSessionInUrl: false,
+    },
+  }
+);
+
+// Auto-refresh pour mobile
+if (Platform.OS !== "web") {
+  AppState.addEventListener("change", (state) => {
+    if (state === "active") {
+      supabase.auth.startAutoRefresh();
+    } else {
+      supabase.auth.stopAutoRefresh();
+    }
+  });
+}
+
+// =====================================================
+// CONFIGURATION LEGEND-STATE
+// =====================================================
+
+export const generateId = () => uuidv4();
+
+// Configuration de base pour toutes les tables
+export const customSynced = configureSynced(syncedSupabase, {
+  persist: {
+    plugin: observablePersistAsyncStorage({
+      AsyncStorage: Storage,
+    }),
+  },
+  generateId,
+  supabase,
+  changesSince: "last-sync",
+  fieldCreatedAt: "created_at",
+  fieldUpdatedAt: "updated_at",
+  fieldDeleted: "deleted",
+});
+
+// Initialiser l'authentification au démarrage
+import { initializeAuth } from "./auth/auth-state";
+initializeAuth();
