+++ utils/supabase/observables/alerts.ts
@@ -1,0 +1,51 @@
+import { observable } from "@legendapp/state";
+import { supabase, customSynced, generateId } from "../client";
+import { Database } from "@/utils/database.types";
+
+// =====================================================
+// OBSERVABLE ALERTS
+// =====================================================
+
+// 🚨 ALERTS
+export const alerts$ = observable(
+  customSynced({
+    supabase,
+    collection: "alerts",
+    select: (from) =>
+      from.select(
+        "id,counter,senior_id,call_id,alert_type,severity,title,description,detected_indicators,confidence_score,status,acknowledged_by,acknowledged_at,resolved_at,resolution_notes,created_at,updated_at,deleted"
+      ),
+    actions: ["read", "create", "update", "delete"],
+    realtime: true,
+    persist: {
+      name: "alerts",
+      retrySync: true,
+    },
+    retry: {
+      infinite: true,
+    },
+  })
+);
+
+// =====================================================
+// HELPER FUNCTIONS
+// =====================================================
+
+export function createAlert(
+  alertData: Partial<Database["public"]["Tables"]["alerts"]["Insert"]>
+) {
+  const id = generateId();
+  alerts$[id].assign({
+    id,
+    status: "new",
+    ...alertData,
+  });
+}
+
+export function acknowledgeAlert(alertId: string, userId: string) {
+  alerts$[alertId].assign({
+    status: "acknowledged",
+    acknowledged_by: userId,
+    acknowledged_at: new Date().toISOString(),
+  });
+}
