+++ utils/supabase/observables/calls.ts
@@ -1,0 +1,27 @@
+import { observable } from "@legendapp/state";
+import { supabase, customSynced } from "../client";
+
+// =====================================================
+// OBSERVABLE CALLS
+// =====================================================
+
+// 📞 CALLS
+export const calls$ = observable(
+  customSynced({
+    supabase,
+    collection: "calls",
+    select: (from) =>
+      from.select(
+        "id,counter,senior_id,call_type,status,started_at,ended_at,duration_seconds,quality_score,conversation_summary,mood_detected,created_at,updated_at,deleted"
+      ),
+    actions: ["read", "create", "update", "delete"],
+    realtime: true,
+    persist: {
+      name: "calls",
+      retrySync: true,
+    },
+    retry: {
+      infinite: true,
+    },
+  })
+);
