+++ utils/supabase/observables/seniors.ts
@@ -1,0 +1,27 @@
+import { observable } from "@legendapp/state";
+import { supabase, customSynced } from "../client";
+
+// =====================================================
+// OBSERVABLE SENIORS
+// =====================================================
+
+// 👴 SENIORS
+export const seniors$ = observable(
+  customSynced({
+    supabase,
+    collection: "seniors",
+    select: (from) =>
+      from.select(
+        "id,counter,user_id,birth_date,preferred_call_time,call_frequency,personality_profile,medical_context,interests,communication_preferences,emergency_contact,address,created_at,updated_at,deleted"
+      ),
+    actions: ["read", "create", "update", "delete"],
+    realtime: true,
+    persist: {
+      name: "seniors",
+      retrySync: true,
+    },
+    retry: {
+      infinite: true,
+    },
+  })
+);
