+++ utils/supabase/observables/todos.ts
@@ -1,0 +1,42 @@
+import { observable } from "@legendapp/state";
+import { supabase, customSynced, generateId } from "../client";
+
+// =====================================================
+// OBSERVABLE TODOS (pour les tests)
+// =====================================================
+
+// Garder les todos pour les tests
+export const todos$ = observable(
+  customSynced({
+    supabase,
+    collection: "todos",
+    select: (from) =>
+      from.select("id,counter,text,done,created_at,updated_at,deleted"),
+    actions: ["read", "create", "update", "delete"],
+    realtime: true,
+    persist: {
+      name: "todos",
+      retrySync: true,
+    },
+    retry: {
+      infinite: true,
+    },
+  })
+);
+
+// =====================================================
+// HELPER FUNCTIONS
+// =====================================================
+
+// Fonctions todos (pour les tests)
+export function addTodo(text: string) {
+  const id = generateId();
+  todos$[id].assign({
+    id,
+    text,
+  });
+}
+
+export function toggleDone(id: string) {
+  todos$[id].done.set((prev: boolean) => !prev);
+}
