+++ utils/supabase/observables/users.ts
@@ -1,0 +1,27 @@
+import { observable } from "@legendapp/state";
+import { supabase, customSynced } from "../client";
+
+// =====================================================
+// OBSERVABLE USERS
+// =====================================================
+
+// 👥 USERS
+export const users$ = observable(
+  customSynced({
+    supabase,
+    collection: "users",
+    select: (from) =>
+      from.select(
+        "id,counter,email,user_type,first_name,last_name,phone,created_at,updated_at,deleted,is_active"
+      ),
+    actions: ["read", "create", "update", "delete"],
+    realtime: true,
+    persist: {
+      name: "users",
+      retrySync: true,
+    },
+    retry: {
+      infinite: true,
+    },
+  })
+);
